+++ src/meili/routing.cc
@@ -166,8 +166,9 @@ inline bool IsEdgeAllowed(const baldr::DirectedEdge* edge,
                           const Label& pred_edgelabel,
                           const baldr::GraphTile* tile,
                           int& restriction_idx) {
-  bool valid_pred = (!pred_edgelabel.edgeid().Is_Valid() && costing->GetEdgeFilter()(edge) != 0.f) ||
-                    edgeid == pred_edgelabel.edgeid();
+  bool valid_pred =
+      (!pred_edgelabel.edgeid().Is_Valid() && costing->Filter(edge, edgeid, tile) != 0.f) ||
+      edgeid == pred_edgelabel.edgeid();
   bool restricted = !costing->Allowed(edge, pred_edgelabel, tile, edgeid, 0, 0, restriction_idx);
   return valid_pred || !restricted;
 }