+++ src/odin/maneuversbuilder.cc
@@ -37,6 +37,9 @@ using namespace valhalla::odin;
 
 namespace {
 
+constexpr uint32_t kRelativeStraightTurnDegreeLowerBound = 330;
+constexpr uint32_t kRelativeStraightTurnDegreeUpperBound = 30;
+
 constexpr float kShortForkThreshold = 0.05f; // Kilometers
 
 // Kilometers - picked since the next rounded maneuver announcement will happen
@@ -110,6 +113,9 @@ std::list<Maneuver> ManeuversBuilder::Build() {
 
   // Process the guidance view junctions
   ProcessGuidanceViewJunctions(maneuvers);
+
+  // Update the maneuver placement for internal intersection turns
+  UpdateManeuverPlacementForInternalIntersectionTurns(maneuvers);
 
 #ifdef LOGGING_LEVEL_TRACE
   int final_man_id = 1;
@@ -2969,6 +2975,133 @@ void ManeuversBuilder::SetTraversableOutboundIntersectingEdgeFlags(std::list<Man
       }
     }
   }
+}
+
+void ManeuversBuilder::UpdateManeuverPlacementForInternalIntersectionTurns(
+    std::list<Maneuver>& maneuvers) {
+
+  auto is_turn_maneuver = [](DirectionsLeg_Maneuver_Type maneuver_type) -> bool {
+    switch (maneuver_type) {
+      case DirectionsLeg_Maneuver_Type_kSlightRight:
+      case DirectionsLeg_Maneuver_Type_kRight:
+      case DirectionsLeg_Maneuver_Type_kSharpRight:
+      case DirectionsLeg_Maneuver_Type_kUturnRight:
+      case DirectionsLeg_Maneuver_Type_kUturnLeft:
+      case DirectionsLeg_Maneuver_Type_kSharpLeft:
+      case DirectionsLeg_Maneuver_Type_kLeft:
+      case DirectionsLeg_Maneuver_Type_kSlightLeft:
+      case DirectionsLeg_Maneuver_Type_kRampRight:
+      case DirectionsLeg_Maneuver_Type_kRampLeft:
+      case DirectionsLeg_Maneuver_Type_kStayRight:
+      case DirectionsLeg_Maneuver_Type_kStayLeft:
+        return true;
+      default:
+        return false;
+    }
+  };
+
+  auto is_relative_straight = [](uint32_t turn_degree) -> bool {
+    return ((turn_degree >= kRelativeStraightTurnDegreeLowerBound) ||
+            (turn_degree <= kRelativeStraightTurnDegreeUpperBound));
+  };
+
+  Maneuver* prev_maneuver = nullptr;
+  for (auto& maneuver : maneuvers) {
+    // Verify that the previous maneuver is set
+    if (prev_maneuver) {
+      // Skip destination maneuver
+      if (maneuver.IsDestinationType()) {
+        break;
+      }
+
+      // Maneuver must be a turn
+      if (is_turn_maneuver(maneuver.type())) {
+        auto original_maneuver_end_node_index = maneuver.end_node_index();
+
+        // Iterate over edges
+        for (auto node_index = maneuver.begin_node_index();
+             node_index < original_maneuver_end_node_index; ++node_index) {
+          auto new_node_index = node_index + 1;
+          auto prev_edge = trip_path_->GetPrevEdge(node_index);
+          auto edge = trip_path_->GetCurrEdge(node_index);
+
+          // New node index must less than original maneuver end node index (cannot move all edges)
+          // maneuvers must be the same travel mode and
+          // edge must be internal and straight from previous maneuver
+          if ((new_node_index < original_maneuver_end_node_index) &&
+              (prev_maneuver->travel_mode() == maneuver.travel_mode()) &&
+              edge->internal_intersection() &&
+              is_relative_straight(GetTurnDegree(prev_edge->end_heading(), edge->begin_heading()))) {
+            // Add straight internal edge to previous maneuver
+            MoveInternalEdgeToPreviousManeuver(*prev_maneuver, maneuver, new_node_index,
+                                               prev_edge.get(), edge.get());
+          } else {
+            // Exit form the edge loop
+            break;
+          }
+        } // end edge loop
+      }
+    }
+
+    // Update previous maneuver
+    prev_maneuver = &maneuver;
+
+  } // end maneuver loop
+}
+
+void ManeuversBuilder::MoveInternalEdgeToPreviousManeuver(Maneuver& prev_maneuver,
+                                                          Maneuver& maneuver,
+                                                          uint32_t new_node_index,
+                                                          EnhancedTripLeg_Edge* prev_edge,
+                                                          EnhancedTripLeg_Edge* edge) {
+  /////////////////////////////////////////////////////////////////////////////
+  // Update the previous maneuver
+
+  // Increase distance in kilometers
+  prev_maneuver.set_length(prev_maneuver.length() + edge->length());
+
+  // Increase basic time (len/speed on each edge with no stop impact) in seconds
+  prev_maneuver.set_basic_time(
+      prev_maneuver.basic_time() +
+      GetTime(edge->length(), GetSpeed(prev_maneuver.travel_mode(), edge->default_speed())));
+
+  // Set the end node index
+  prev_maneuver.set_end_node_index(new_node_index);
+
+  // Set the end shape index
+  prev_maneuver.set_end_shape_index(edge->end_shape_index());
+
+  // Update the time based on the delta of the elapsed time between the begin
+  // and end nodes
+  prev_maneuver.set_time(
+      trip_path_->node(prev_maneuver.end_node_index()).cost().elapsed_cost().seconds() -
+      trip_path_->node(prev_maneuver.begin_node_index()).cost().elapsed_cost().seconds());
+
+  /////////////////////////////////////////////////////////////////////////////
+  // Update the maneuver
+
+  // Decrease distance in kilometers
+  maneuver.set_length(maneuver.length() - edge->length());
+
+  // Decrease basic time (len/speed on each edge with no stop impact) in seconds
+  maneuver.set_basic_time(
+      maneuver.basic_time() -
+      GetTime(edge->length(), GetSpeed(maneuver.travel_mode(), edge->default_speed())));
+
+  // Set the begin node index
+  maneuver.set_begin_node_index(new_node_index);
+
+  // Set the begin shape index
+  maneuver.set_begin_shape_index(edge->end_shape_index());
+
+  // Update the time based on the delta of the elapsed time between the begin
+  // and end nodes
+  maneuver.set_time(trip_path_->node(maneuver.end_node_index()).cost().elapsed_cost().seconds() -
+                    trip_path_->node(maneuver.begin_node_index()).cost().elapsed_cost().seconds());
+
+  /////////////////////////////////////////////////////////////////////////////
+  // Update the edge turn lanes with the previous edge turn lanes
+  edge->mutable_turn_lanes()->CopyFrom(prev_edge->turn_lanes());
 }
 
 } // namespace odin