+++ src/thor/attributes_controller.cc
@@ -73,6 +73,7 @@ const std::unordered_map<std::string, bool> AttributesController::kDefaultAttrib
     {kEdgeDefaultSpeed, true},
     {kEdgeDestinationOnly, true},
     {kEdgeIsUrban, false},
+    {kEdgeTaggedNames, true},
 
     // Node keys
     {kNodeIncidents, false},