+++ src/tyr/route_serializer_osrm.cc
@@ -546,6 +546,17 @@ json::ArrayPtr intersections(const valhalla::DirectionsLeg::Maneuver& maneuver,
     intersection->emplace("entry", entries);
     intersection->emplace("bearings", bearings);
 
+    // Add tunnel_name for tunnels
+    if (!arrive_maneuver) {
+      if (curr_edge->tunnel() && !curr_edge->tagged_name().empty()) {
+        for (uint32_t t = 0; t < curr_edge->tagged_name().size(); ++t) {
+          if (curr_edge->tagged_name().Get(t).type() == TaggedName_Type_kTunnel) {
+            intersection->emplace("tunnel_name", curr_edge->tagged_name().Get(t).value());
+          }
+        }
+      }
+    }
+
     // Add classes based on the first edge after the maneuver (not needed
     // for arrive maneuver).
     if (!arrive_maneuver) {
@@ -563,10 +574,9 @@ json::ArrayPtr intersections(const valhalla::DirectionsLeg::Maneuver& maneuver,
         classes.push_back("ferry");
       }
 
-      /** TODO
-      if ( ) {
+      if (curr_edge->destination_only()) {
         classes.push_back("restricted");
-      } */
+      }
       if (classes.size() > 0) {
         auto class_list = json::array({});
         for (const auto& cl : classes) {